Reformat with Rubocop
A consistent style helps when editing the code. Without any strong
opinions I largely went for the rubocop defaults, with automatic
corrections doing the bulk of the work.

Some method variable names were changed to longer ones. There
is a Cop that identifies methods with short names. These changes
were not just to satisfy the cop, as they also make the code
easier to read.